webserver/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
edge-tts==7.0.2
Flask[async]==3.1.1
flask-cors==6.0.1
flask-socketio==4.3.2 # Pinned for compatibility with esp libraries
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
multidict==6.6.3
openai==1.97.1
pillow==11.3.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.1
# python-socketio==4.3.0
requests==2.32.4
sniffio==1.3.1
srt==3.5.3
tabulate==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
vosk==0.3.45
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
gunicorn
eventlet